pyproject.toml
[project]
name = "mattermost-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = ["requests>=2.32.3", "python-dotenv>=1.0.1", "openai>=1.59.6"]
license = { text = "MIT" }
